ci: test and release fixes
