+++ test/secret-config.yml
@@ -2,6 +2,6 @@ hierarchy:
   - name: test
     namespace: default
     kind: Secret
-  - name: "spring-defaults-{{.}}"
+  - name: "spring-defaults-{{.name}}"
     namespace: default
     kind: ConfigMap
